+++ pandas/tests/indexes/period/test_setops.py
@@ -4,7 +4,7 @@ import pytest
 from pandas._libs.tslibs import IncompatibleFrequency
 
 import pandas as pd
-from pandas import Index, PeriodIndex, date_range, period_range
+from pandas import PeriodIndex, date_range, period_range
 import pandas._testing as tm
 
 
@@ -13,34 +13,6 @@ def _permute(obj):
 
 
 class TestPeriodIndex:
-    def test_joins(self, join_type):
-        index = period_range("1/1/2000", "1/20/2000", freq="D")
-
-        joined = index.join(index[:-5], how=join_type)
-
-        assert isinstance(joined, PeriodIndex)
-        assert joined.freq == index.freq
-
-    def test_join_self(self, join_type):
-        index = period_range("1/1/2000", "1/20/2000", freq="D")
-
-        res = index.join(index, how=join_type)
-        assert index is res
-
-    def test_join_does_not_recur(self):
-        df = tm.makeCustomDataframe(
-            3,
-            2,
-            data_gen_f=lambda *args: np.random.randint(2),
-            c_idx_type="p",
-            r_idx_type="dt",
-        )
-        s = df.iloc[:2, 0]
-
-        res = s.index.join(df.columns, how="outer")
-        expected = Index([s.index[0], s.index[1], df.columns[0], df.columns[1]], object)
-        tm.assert_index_equal(res, expected)
-
     @pytest.mark.parametrize("sort", [None, False])
     def test_union(self, sort):
         # union
@@ -180,10 +152,6 @@ class TestPeriodIndex:
         index2 = period_range("1/1/2000", "1/20/2000", freq="W-WED")
         with pytest.raises(IncompatibleFrequency):
             index.union(index2, sort=sort)
-
-        index3 = period_range("1/1/2000", "1/20/2000", freq="2D")
-        with pytest.raises(IncompatibleFrequency):
-            index.join(index3)
 
     # TODO: belongs elsewhere
     def test_union_dataframe_index(self):
